+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `catalystcenter_sda_fabric_site` resource and data source
 - Add `catalystcenter_template_version` resource and data source
 - Add `catalystcenter_sda_fabric_authentication_profile` resource and data source
+- Fix handling of validation warnings when creating templates
 
 ## 0.1.4